tests: split conftest.py into tests helpers
